+++ src/plaid_bridges/common/__init__.py
@@ -1,9 +1,9 @@
 """Module that implements PLAID bridges."""
 
-from plaid_bridges.common.base_regression_dataset import (
+from plaid_bridges.common.base_regression import (
     BaseRegressionDataset,
     BaseTransformer,
 )
-from plaid_bridges.common.homogeneous_dataset import HomogeneousOfflineTransformer
+from plaid_bridges.common.homogeneous import HomogeneousOfflineTransformer
 
 __all__ = ["BaseRegressionDataset", "BaseTransformer", "HomogeneousOfflineTransformer"]
